feat: Consume shared components from Maven
*   Adds a `use_local_shared_components` flag to `gradle.properties`.
*   Conditionally adds `:shared:components` locally or from Maven based on the flag.
*   Removes the local `:shared:components` module dependency.
*   Moves navigation and presentation logic from `:shared:navigation` to the `:composeApp` module.
