+++ build-logic/convention/src/main/kotlin/AppIdentity.kt
@@ -3,7 +3,8 @@ import org.gradle.api.Project
 data class AppIdentity(
     val appName: String,
     val packagePrefix: String,
-    val mainPackageName: String
+    val mainPackageName: String,
+    val localComponents: Boolean
 ) {
     val packageName: String
         get() = "$packagePrefix.$mainPackageName"
@@ -16,5 +17,6 @@ val Project.appIdentity: AppIdentity
         packagePrefix = findProperty("appPackagePrefix")?.toString()
             ?: error("APP_PACKAGE_PREFIX not defined in gradle.properties"),
         mainPackageName = findProperty("appPackageName")?.toString()
-            ?: error("MAIN_PACKAGE_NAME not defined in gradle.properties")
+            ?: error("MAIN_PACKAGE_NAME not defined in gradle.properties"),
+        localComponents = findProperty("use_local_shared_components")?.toString()?.toBoolean() ?: false
     )
