+++ build-logic/convention/src/main/kotlin/ke/don/ski/extensions/ModulesExtension.kt
@@ -48,8 +48,7 @@ val Project.sharedModules: SharedModules
 class SharedModules(private val project: Project) {
     val design get() = project.project(":shared:design")
     val resources get() = project.project(":shared:resources")
-    val common get() = project.project(":shared:components")
     val navigation get() = project.project(":shared:navigation")
 
-    val all get() = listOf(design, resources, common, navigation)
+    val all get() = listOf(design, resources, navigation)
 }
