+++ composeApp/src/commonMain/kotlin/ke/don/ski/presentation/ScaffoldComponents.kt
@@ -1,4 +1,4 @@
-package ke.don.navigation
+package ke.don.ski.presentation
 
 import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.layout.Arrangement
