+++ gradle.properties
@@ -15,3 +15,4 @@ android.useAndroidX=true
 appPackagePrefix=ke.don
 appPackageName=ski
 appName=ski
+use_local_shared_components=false
