+++ gradle/libs.versions.toml
@@ -16,7 +16,8 @@ kotlin = "2.3.0"
 kotlinx-coroutines = "1.10.2"
 compose = "1.10.0-rc02"
 kotlinxSerializationCore = "1.9.0"
-vanniktech = "0.34.0"
+vanniktech = "0.35.0"
+ski = "v0.0.1-beta2"
 
 [libraries]
 kotlin-test = { module = "org.jetbrains.kotlin:kotlin-test", version.ref = "kotlin" }
@@ -30,6 +31,8 @@ androidx-activity-compose = { module = "androidx.activity:activity-compose", ver
 androidx-lifecycle-viewmodelCompose = { module = "org.jetbrains.androidx.lifecycle:lifecycle-viewmodel-compose", version.ref = "androidx-lifecycle" }
 androidx-lifecycle-runtimeCompose = { module = "org.jetbrains.androidx.lifecycle:lifecycle-runtime-compose", version.ref = "androidx-lifecycle" }
 kotlinx-coroutinesSwing = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-swing", version.ref = "kotlinx-coroutines" }
+
+ski-components = {module = "io.github.donald-okara:ski", version.ref = "ski"}
 
 # Serialization
 kotlinx-serialization-core = { module = "org.jetbrains.kotlinx:kotlinx-serialization-core", version.ref = "kotlinxSerializationCore" }
